backend/requirements.txt
fastapi
uvicorn
supabase
httpx
python-dotenv
pydantic
Pillow
